+++ test/IntegrationTests.js
@@ -498,4 +498,25 @@ describe('BitX', function() {
       });
     });
   });
+
+  describe('getWithdrawal', function() {
+
+    it('should return the withdrawal', function(done) {
+      var expectedWithdrawal = {
+        status: 'COMPLETED',
+        id: '1212'
+      };
+
+      server.on('request', function(req, res) {
+        expect(req).to.have.property('method', 'GET');
+        expect(req).to.have.property('url', '/api/1/withdrawals/1212');
+        res.end(JSON.stringify(expectedWithdrawal));
+      });
+
+      bitx.getWithdrawal('1212', function(err, withdrawal) {
+        expect(withdrawal).to.eql(expectedWithdrawal);
+        done(err);
+      });
+    });
+  });
 });